Create file board.rb
